refactor: rename view directory to views
